apps/video-analyzer-api/requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.5
django-3-jet==1.0.8
django-colorfield==0.14.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-resized==1.0.3
djangorestframework==3.16.1
djangorestframework-jsonapi==8.0.0
djangorestframework_simplejwt==5.5.1
drf-extra-fields==3.7.0
drf-yasg==1.21.10
environ==1.0
filetype==1.2.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsmin==3.0.1
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
pycparser==2.22
pyenchant==3.2.2
PyJWT==2.10.1
pyparsing==3.2.3
python-environ==0.4.54
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
